requirements.txt
Flask
gunicorn
python-dotenv
requests
openai
google-cloud-aiplatform
groq
